backend/requirements.txt
# Core LLM & RAG Dependencies
anthropic>=0.39.0
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging & Monitoring
loguru>=0.7.2
